Modified the TypeORM entities such that custom join tables can be used with whatever extra columns are necessary. It is also possible to easily join from either side of the relation.
